XML schema for fixtures & some fixture fixes

+++ engine/src/qlcphysical.h
@@ -101,11 +101,11 @@ public:
     void setLensName(const QString& name);
     QString lensName() const;
 
-    void setLensDegreesMin(int degrees);
-    int lensDegreesMin() const;
+    void setLensDegreesMin(qreal degrees);
+    qreal lensDegreesMin() const;
 
-    void setLensDegreesMax(int degrees);
-    int lensDegreesMax() const;
+    void setLensDegreesMax(qreal degrees);
+    qreal lensDegreesMax() const;
 
     void setFocusType(const QString& type);
     QString focusType() const;
@@ -133,8 +133,8 @@ protected:
     int m_depth;
 
     QString m_lensName;
-    int m_lensDegreesMin;
-    int m_lensDegreesMax;
+    qreal m_lensDegreesMin;
+    qreal m_lensDegreesMax;
 
     QString m_focusType;
     int m_focusPanMax;
